requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.1.0
faiss-cpu>=1.7.0  # Use faiss-gpu if GPU is available

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0

# Security
bcrypt>=3.2.0
PyJWT>=2.1.0
cryptography>=3.4.7

# Database
SQLAlchemy>=1.4.23

# Monitoring and Logging
prometheus-client>=0.11.0
python-json-logger>=2.0.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3
pre-commit>=2.15.0

tqdm>=4.62.0
plotly>=5.3.1
matplotlib>=3.4.3
seaborn>=0.11.2
jupyter>=1.0.0
notebook>=6.4.5 